Add COinS metadata parser
Add parseCOinS which selects all span tags with
class Z3988, and parses the title tag into an
object. Hierarchical keys (e.g. rft.date) are
split, with the top level key rft pointing to
an object with the second level key as a key
in that object. Then the object is added to the
list, as multiple COinS objects can exist in
a page.

There is also a separate exportable parsing function
for the contents of the title tag, which can be
used separately, i.e. for parsing the contents of
a cross ref request i.e.
http://search.crossref.org/dois?q=10.5555%2F12345678

Phabricator issue: https://phabricator.wikimedia.org/T104174

Bug: T104174