Fix tests
Fix two tests that were failing due to changes
in the metadata available upstream.

Fix additional two tests which resulted in an
uncaught error instead of failing the test; and
added headers for woorank which now returns
a 500 internal server error if a request
is made without a user-agent header.