[BREAKING CHANGE] Remove preq from library
Remove preq from library entirely and
replace with native fetch.

This is not backwards compatible if third
party users are a) using the parseAll function,
and b) other features of preq, such as
the internalUri parameter or wrapped errors;
see https://github.com/wikimedia/preq for more
details on how preq alters the Response
object.

Bug: T361602