Remove support for callback (node-style) requests
Bluebird Promises provides support for nodeifying this
library. However bluebird use is now discouraged. In
preperation for removing this library and switching to native
JS Promises, remove support for callback style requests.

Increment package.json to 3.0.0 to go along with documentation, but
not ready for release yet, as this release is planned
to remove the bluebird library entirely.